api/requirements.txt
azure-functions  
requests  
python-docx  